using cJSON instead of homemade solution

+++ main/weather.h
@@ -7,7 +7,7 @@
 #define API_CALL "http://api.openweathermap.org/data/2.5/weather?id=%s&units=metric&APPID=%s"
 #define MALMOE_ID "2692969"
 
-float get_temperature();
+double get_temperature();
 
 void update_weather_information(EventGroupHandle_t wifi_event_group);
